% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_bonjour.R
\name{dire_bonjour}
\alias{dire_bonjour}
\title{Fonction pour dire bonjour}
\usage{
dire_bonjour(prenom = "toi")
}
\arguments{
\item{prenom}{Le prénom de la personne au format 'string'. Par défaut, le prénom est 'toi'.}
}
\value{
print 'Bonjour, <prénom>'
}
\description{
Cette fonction dit bonjour à une personne qui lui donne son prénom.
}
\examples{
dire_bonjour("Cindie")
dire_bonjour()
}
